requirements.txt
fastapi
uvicorn
python-telegram-bot
openai
google-api-python-client
google-auth
python-docx  # якщо хочеш працювати з docx шаблонами